+++ components/header/dateRender.js
@@ -1,6 +1,6 @@
-import { sharedState } from "../state/state.js";
-import { loadEntriesFromServer } from "../api/api.js";
-import { getDateFromServer } from "./entry.js";
+import { sharedState } from "../../../store/state.js";
+import { loadEntriesFromServer } from "../../../api.js";
+import { getDateFromServer } from "../main/ledger/entries/entry.js";
 
 /*
   여기서 year과 month를 store로 관리하는 것이 좋을 것 같습니다.
@@ -9,7 +9,7 @@ import { getDateFromServer } from "./entry.js";
 export let currentMonth = Date.now() ? new Date().getMonth() + 1 : 1; // 현재 월 (1~12)
 export let currentYear = Date.now() ? new Date().getFullYear() : 2023; // 현재 연도
 
-export function initCalendar() {
+export function initDate() {
   const yearEl = document.getElementById("year");
   const monthEl = document.getElementById("month");
   const monthLabelEl = document.getElementById("month-label");
