+++ components/main/ledger/entries/filterEntries.js
@@ -1,6 +1,6 @@
 // 필터링 기능 구현
-import { sharedState } from "../state/state.js";
-import { updateTotalAmounts } from "./totalAmount.js";
+import { sharedState } from "../../../../store/state.js";
+import { updateTotalAmounts } from "../totalAmount/totalAmount.js";
 import { updateDateSectionTotals } from "./entry.js";
 
 export function initFilterButtons() {
