--- js/dummy/dummy.js
@@ -1,29 +1,0 @@
-/*
-    이제 서버에서 가져오므로 사용안함
-*/
-export const dummyEntries = [
-  {
-    date: "2025-07-10",
-    amount: 50000,
-    desc: "월급",
-    method: "카드",
-    category: "문화/여가",
-    isIncome: true
-  },
-  {
-    date: "2025-07-11",
-    amount: 15000,
-    desc: "점심",
-    method: "현금",
-    category: "용돈",
-    isIncome: false
-  },
-  {
-    date: "2025-07-12",
-    amount: 30000,
-    desc: "영화 관람",
-    method: "카드",
-    category: "식비",
-    isIncome: false
-  }
-];
