--- js/dummy/message.js
@@ -1,3 +1,0 @@
-export const hello = () =>{
-    return "hello world";
-}
